python_service/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pillow==10.1.0
python-dotenv==1.0.1
requests==2.31.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.0+cpu
torchvision==0.17.0+cpu
numpy==1.26.4 